Refactored AndroidBLEAdvanced to use correct package names for the Makeability Lab
