+++ .gitignore
@@ -67,3 +67,5 @@ A03-BLEAdvanced/AndroidBLEAdvanced/.idea/caches/build_file_checksums.ser
 L07-OpenCV/OpenCVImageProcessing/.idea/caches/build_file_checksums.ser
 *.ser
 *.ser
+*.ser
+*.ser
