+++ A03-BLEAdvanced/AndroidBLEAdvanced/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 26
     defaultConfig {
-        applicationId "com.example.lianghe.pincontroldemo_android"
+        applicationId "io.makeabilitylab.bledemo"
         minSdkVersion 21
         targetSdkVersion 26
         versionCode 1
