+++ A03-BLEAdvanced/AndroidBLEAdvanced/app/src/main/java/io/makeabilitylab/bledemo/ble/BLEDevice.java
@@ -1,4 +1,4 @@
-package com.example.lianghe.android_ble_advanced.ble;
+package io.makeabilitylab.bledemo.ble;
 
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
