+++ A03-BLEAdvanced/AndroidBLEAdvanced/app/src/main/java/io/makeabilitylab/bledemo/ble/BLEListener.java
@@ -1,4 +1,4 @@
-package com.example.lianghe.android_ble_advanced.ble;
+package io.makeabilitylab.bledemo.ble;
 
 /** Listens to events related to the device. */
 public interface BLEListener {
